Re-structuring numpy indexing routine functions to their respective files
